requirements.txt
Flask>=2.2.2
numpy>=1.24.3
pandas>=2.0.3
pillow>=10.2.0
python-docx>=1.1.0
requests>=2.31.0